requirements.txt
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
google-auth==2.38.0
google-auth-oauthlib==1.2.1
gspread==6.2.0
h11==0.14.0
httpcore==1.0.8
httpx==0.25.2
idna==3.10
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.1.0
python-telegram-bot==20.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sniffio==1.3.1
typing_extensions==4.13.2
urllib3==2.4.0
